requirements.txt
alabaster==0.7.10
Babel==2.3.4
click==6.7
docutils==0.13.1
Flask==0.12
Flask_API==1.0
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
pudb==2017.1.1
py==1.4.32
pycrypto==2.6.1
Pygments==2.2.0
pytest==3.0.6
pytz==2016.10
requests==2.13.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.7.1
sphinx-autodoc-typehints==1.2.5
treelib==1.3.5
urwid==1.3.1
Werkzeug==0.12
